# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_pozgo_web-cli_internal_models.Server:
    properties:
      created_at:
        type: string
      group:
        description: Group/category for organization
        type: string
      id:
        type: integer
      ip_address:
        description: IP address
        type: string
      name:
        description: Hostname (must follow hostname conventions)
        type: string
      port:
        description: 'SSH port (default: 22)'
        type: integer
      source:
        description: '"sqlite" or "vault"'
        type: string
      updated_at:
        type: string
      username:
        description: SSH username for remote connections
        type: string
    type: object
  github_com_pozgo_web-cli_internal_models.ServerCreate:
    properties:
      group:
        description: Optional, defaults to "default"
        type: string
      ip_address:
        type: string
      name:
        type: string
      port:
        description: Optional, defaults to 22 if not provided
        type: integer
      username:
        description: SSH username for remote connections
        type: string
    type: object
  github_com_pozgo_web-cli_internal_models.ServerUpdate:
    properties:
      group:
        type: string
      ip_address:
        type: string
      name:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
...
  internal_server.ErrorResponse:
    description: Error response returned by the API
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
host: localhost:7777
info:
  contact:
    email: linux@ozgo.info
    name: Web CLI Support
    url: https://github.com/pozgo/web-cli
  description: Basic authentication using username and password
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Web CLI API
  version: 1.1.0
paths:
...
  /servers:
    get:
      consumes:
      - application/json
      description: Get a list of all remote servers configured in the system
      parameters:
      - description: Filter by group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.Server'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all servers
      tags:
      - Servers
    post:
      consumes:
      - application/json
      description: Add a new remote server configuration
      parameters:
      - description: Server to create
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ServerCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create a new server
      tags:
      - Servers
  /servers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a server configuration by its ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete a server
      tags:
      - Servers
    get:
      consumes:
      - application/json
      description: Get a single server configuration by its ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get a server by ID
      tags:
      - Servers
    put:
      consumes:
      - application/json
      description: Update an existing server configuration by its ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Server update data
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.ServerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update a server
      tags:
      - Servers
...
  /vault/bash-scripts:
    get:
      description: Retrieve all bash scripts stored in Vault
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_vault.BashScript'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List bash scripts from Vault
      tags:
      - Vault
    post:
      consumes:
      - application/json
      description: Create a new bash script in HashiCorp Vault storage
      parameters:
      - description: Bash script data
        in: body
        name: script
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created script info with source field
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or Vault not configured
          schema:
            type: string
        "500":
          description: Failed to store script
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Store bash script in Vault
      tags:
      - Vault
  /vault/config:
    delete:
      description: Remove the Vault configuration
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete Vault configuration
      tags:
      - Vault
    get:
      description: Retrieve the current Vault configuration (token is never returned)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.VaultConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get Vault configuration
      tags:
      - Vault
    post:
      consumes:
      - application/json
      description: Configure or update the Vault connection settings
      parameters:
      - description: Vault configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.VaultConfigCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.VaultConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create or update Vault configuration
      tags:
      - Vault
  /vault/env-variables:
    get:
      description: Retrieve all environment variables stored in Vault
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pozgo_web-cli_internal_vault.EnvVariable'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List environment variables from Vault
      tags:
      - Vault
    post:
      consumes:
      - application/json
      description: Store a new environment variable in Vault
      parameters:
      - description: Environment Variable
        in: body
        name: envVar
        required: true
        schema:
          properties:
            description:
              type: string
            group:
              type: string
            name:
              type: string
            value:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              description:
                type: string
              group:
                type: string
              name:
                type: string
              source:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create environment variable in Vault
      tags:
      - Vault
...
  /vault/status:
    get:
      description: Check the current Vault connection status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.VaultStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get Vault connection status
      tags:
      - Vault
  /vault/test:
    post:
      description: Test the connection to Vault and initialize the secrets structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pozgo_web-cli_internal_models.VaultStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Test Vault connection
      tags:
      - Vault
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...